* Perform all vector math for bubbles with single precision floating point numbers to reduce computation overhead   - Especially since most OpenGL implementations implicitly convert to 32 bit or less floats anyway.
git-svn-id: https://svn.il.fontys.nl/usersvn/fishtank/svn/trunk@286 bb1b31de-baa9-489e-af1c-03370f2903b8

+++ fishtank/AquariumController.cpp
@@ -55,7 +55,7 @@ void AquariumController::AddObject(boost::shared_ptr<Model> model, const string 
 	objects.push_back(Object(model, propertiesFile, position));
 }
 
-void AquariumController::AddBubbleSpot(const Eigen::Vector3d &position)
+void AquariumController::AddBubbleSpot(const Eigen::Vector3f& position)
 {
 	bubbleSpots.push_back(position);
 }
@@ -67,7 +67,7 @@ void AquariumController::Update(double dt)
 	foreach (Object& object, objects)
 		object.Update(dt);
 	//voeg op willekeurige momenten bubbels toe
-	foreach (const Eigen::Vector3d& bubbleSpot, bubbleSpots)
+	foreach (const Eigen::Vector3f& bubbleSpot, bubbleSpots)
 	{
 		if (my_random() < dt * 9.5)
 		{
